% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyNumbers.R
\name{asDouble}
\alias{asDouble}
\alias{asDouble,lazyVector,numeric-method}
\alias{asDouble,lazyMatrix,numeric-method}
\alias{asDouble,lazyVector,missing-method}
\alias{asDouble,lazyMatrix,missing-method}
\title{Coerce lazy numbers to double numbers}
\usage{
\S4method{asDouble}{lazyVector,numeric}(x, prec = 1e-15)

\S4method{asDouble}{lazyVector,missing}(x, prec)

\S4method{asDouble}{lazyMatrix,numeric}(x, prec = 1e-15)

\S4method{asDouble}{lazyMatrix,missing}(x, prec)
}
\arguments{
\item{x}{a lazy vector or a lazy matrix}

\item{prec}{relative precision, a number between 0 and 1}
}
\value{
A numeric vector or a numeric matrix.
}
\description{
Coerce a lazy vector (\code{lazyVector} object) to a numeric 
  vector and a lazy matrix (\code{lazyMatrix} object) to a numeric matrix.
}
\note{
Once \code{as.double} or \code{asDouble} (\code{as.double(x)} is 
  equivalent to \code{asDouble(x, prec = 1e-15)}) has been applied to a 
  lazy vector, then this lazy vector is resolved (see 
  \code{\link{lazyResolve}}).
}
\examples{
x <- 1 - lazynb(7) * 0.1
asDouble(x, prec = 1e-15) == 0.3
asDouble(x, prec = 1e-16) == 0.3
}
