% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyNumbers.R
\name{asNumeric}
\alias{asNumeric}
\alias{asNumeric,lazyVector,numeric-method}
\alias{asNumeric,lazyMatrix,numeric-method}
\alias{asNumeric,lazyVector,missing-method}
\alias{asNumeric,lazyMatrix,missing-method}
\title{Coerce lazy numbers to double numbers}
\usage{
\S4method{asNumeric}{lazyVector,numeric}(x, prec = 1e-15)

\S4method{asNumeric}{lazyVector,missing}(x, prec)

\S4method{asNumeric}{lazyMatrix,numeric}(x, prec = 1e-15)

\S4method{asNumeric}{lazyMatrix,missing}(x, prec)
}
\arguments{
\item{x}{a lazy vector or a lazy matrix}

\item{prec}{relative precision, a number between 0 and 1}
}
\value{
A numeric vector or a numeric matrix.
}
\description{
Coerce a lazy vector (\code{lazyVector} object) to a numeric 
  vector and a lazy matrix (\code{lazyMatrix} object) to a numeric matrix.
}
\note{
\code{as.double(x)} is equivalent to \code{asNumeric(x, prec = 1e-15)}.
}
\examples{
x <- 1 - lazynb(7) * 0.1
asNumeric(x, prec = 1e-15) == 0.3
asNumeric(x, prec = 1e-16) == 0.3
}
