% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concat.R
\name{concat-lazyVectors}
\alias{concat-lazyVectors}
\alias{c,lazyVector-method}
\title{Concatenation of lazy vectors}
\usage{
\S4method{c}{lazyVector}(x, ...)
}
\arguments{
\item{x}{a \code{lazyVector} object}

\item{...}{some \code{lazyVector} objects}
}
\description{
Concatenate two or more \code{lazyVector} objects.
}
