% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyMatrices.R
\name{diag}
\alias{diag}
\alias{diag,lazyMatrix-method}
\title{Extract diagonal of a lazy matrix}
\usage{
\S4method{diag}{lazyMatrix}(x)
}
\arguments{
\item{x}{a square lazy matrix}
}
\value{
The diagonal of \code{x} as a lazy vector.
}
\description{
Extract the diagonal of a square lazy matrix.
}
