% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyNumbers.R
\name{intervals}
\alias{intervals}
\title{Intervals for lazy numbers}
\usage{
intervals(x)
}
\arguments{
\item{x}{a \code{lazyVector} object}
}
\value{
A numeric vector of length two if \code{x} contains only one lazy 
  number, otherwise a numeric matrix with two columns and \code{x@length} 
  rows.
}
\description{
For each lazy number in a \code{lazyVector} object, this function 
  returns an interval containing this lazy number.
}
\examples{
library(lazyNumbers)
x <- lazynb(22) / lazynb(7)
print(intervals(x), digits = 17L)
}
