% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyMatrices.R
\name{lazyDet}
\alias{lazyDet}
\title{Determinant of lazy matrix}
\usage{
lazyDet(M)
}
\arguments{
\item{M}{a \code{lazyMatrix} object corresponding to a square matrix}
}
\value{
A lazy number (\code{lazyVector} object with length 1).
}
\description{
Compute the determinant of a lazy matrix.
}
\examples{
M <- lazymat(toeplitz(c(3, 2, 1)))
as.double(lazyDet(M))
}
