% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arith.R
\name{lazyMatrix-product}
\alias{lazyMatrix-product}
\alias{\%*\%,lazyMatrix,lazyMatrix-method}
\alias{\%*\%,lazyMatrix,matrix-method}
\alias{\%*\%,lazyMatrix,numeric-method}
\alias{\%*\%,matrix,lazyMatrix-method}
\alias{\%*\%,numeric,lazyMatrix-method}
\title{Matricial product of lazy matrices}
\usage{
\S4method{\%*\%}{lazyMatrix,lazyMatrix}(x, y)
}
\arguments{
\item{x, y}{objects of class \code{lazyMatrix}}
}
\value{
A \code{lazyMatrix} object.
}
\description{
Matricial product of lazy matrices.
}
