% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arith.R
\name{lazyMatrix-unary}
\alias{lazyMatrix-unary}
\alias{+,lazyMatrix,missing-method}
\alias{-,lazyMatrix,missing-method}
\title{Unary operators for lazy matrices}
\usage{
\S4method{+}{lazyMatrix,missing}(e1, e2)

\S4method{-}{lazyMatrix,missing}(e1, e2)
}
\arguments{
\item{e1}{object of class \code{lazyMatrix}}

\item{e2}{nothing}
}
\value{
A \code{lazyMatrix} object.
}
\description{
Unary operators for lazy matrices.
}
