% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyMatrices.R
\name{lazyMatrix}
\alias{lazyMatrix}
\alias{as.lazyMatrix}
\alias{lazymat}
\title{Lazy matrices}
\usage{
as.lazyMatrix(M)

lazymat(M)
}
\arguments{
\item{M}{a numeric matrix, a numeric vector, or a \code{lazyVector} object}
}
\value{
An object of class \code{lazyMatrix}.
}
\description{
Create a lazy matrix.
}
\examples{
library(lazyNumbers)
M <- lazymat(toeplitz(c(1, 2)))
as.double(M + M)
as.double(M * M)
as.double(M \%*\% M)
}
