% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arith.R
\name{lazyVector-unary}
\alias{lazyVector-unary}
\alias{+,lazyVector,missing-method}
\alias{-,lazyVector,missing-method}
\title{Unary operators for lazy vectors}
\usage{
\S4method{+}{lazyVector,missing}(e1, e2)

\S4method{-}{lazyVector,missing}(e1, e2)
}
\arguments{
\item{e1}{object of class \code{lazyVector}}

\item{e2}{nothing}
}
\value{
A \code{lazyVector} object.
}
\description{
Unary operators for lazy vectors.
}
