% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyNumbers.R
\name{lazyVector}
\alias{lazyVector}
\alias{as.lazyVector}
\alias{as.lazyNumber}
\alias{lazyvec}
\alias{lazynb}
\title{Lazy vector}
\usage{
as.lazyVector(x)

as.lazyNumber(x)

lazyvec(x)

lazynb(x)
}
\arguments{
\item{x}{a numeric vector or a lazy matrix (\code{lazyMatrix} object)}
}
\value{
An object of class \code{lazyVector}.
}
\description{
Create a vector of lazy numbers.
}
\examples{
library(lazyNumbers)
1 - 7 * 0.1 == 0.3 # FALSE
x <- lazynb(1) - lazynb(7) * lazynb(0.1)
as.double(x) == 0.3 # TRUE
}
