% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyNumbers.R
\docType{data}
\name{NA_lazy_}
\alias{NA_lazy_}
\title{The missing lazy value.}
\format{
An object of class \code{lazyVector} of length 1.
}
\usage{
NA_lazy_
}
\description{
The missing lazy value.
}
\keyword{datasets}
