% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract.R
\name{Submatrix}
\alias{Submatrix}
\alias{[,lazyMatrix,numeric-method}
\alias{[,lazyMatrix,numeric,numeric,ANY-method}
\alias{[,lazyMatrix,numeric,numeric-method}
\alias{[,lazyMatrix,numeric,missing-method}
\alias{[,lazyMatrix,missing,numeric-method}
\alias{[,lazyMatrix,numeric,missing,missing-method}
\alias{[,lazyMatrix,numeric,missing,ANY-method}
\alias{[,lazyMatrix,missing,numeric,ANY-method}
\title{Extract lazy submatrix}
\usage{
\S4method{[}{lazyMatrix,numeric,numeric,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{lazyMatrix,numeric,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{lazyMatrix,numeric,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{lazyMatrix,missing,numeric,ANY}(x, i, j, drop)
}
\arguments{
\item{x}{a \code{lazyMatrix} object}

\item{i, j}{indices}

\item{...}{ignored}

\item{drop}{ignored}
}
\value{
A \code{lazyMatrix} object.
}
\description{
Extract a submatrix of a lazy matrix.
}
