% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract.R
\name{Subvector}
\alias{Subvector}
\alias{[,lazyVector,numeric-method}
\alias{[,lazyVector,numeric,ANY,ANY-method}
\alias{[,lazyVector,logical-method}
\alias{[<-,lazyVector,numeric,missing,lazyVector-method}
\alias{[,lazyVector,logical,ANY,ANY-method}
\title{Extract/replace in a lazy vector}
\usage{
\S4method{[}{lazyVector,numeric,ANY,ANY}(x, i)

\S4method{[}{lazyVector,logical,ANY,ANY}(x, i)

\S4method{[}{lazyVector,numeric,missing,lazyVector}(x, i, j) <- value
}
\arguments{
\item{x}{a \code{lazyVector} object}

\item{i}{indices}

\item{j}{nothing}

\item{value}{a \code{lazyVector} object}
}
\value{
A \code{lazyVector} object.
}
\description{
Extract or replace elements in a lazy vector.
}
