% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transpose.R
\name{transpose-lazyMatrix}
\alias{transpose-lazyMatrix}
\alias{t}
\alias{t,lazyMatrix-method}
\alias{t,lazyVector-method}
\title{Transposition of lazy matrices}
\usage{
\S4method{t}{lazyVector}(x)

\S4method{t}{lazyMatrix}(x)
}
\arguments{
\item{x}{a \code{lazyMatrix} or \code{lazyVector} object}
}
\value{
A \code{lazyMatrix} object.
}
\description{
Transpose a \code{lazyMatrix} object.
}
