% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concat.R
\name{bind-lazyMatrices}
\alias{bind-lazyMatrices}
\alias{cbind}
\alias{rbind}
\alias{cbind,lazyMatrix-method}
\alias{rbind,lazyMatrix-method}
\title{Concatenation of lazy matrices}
\usage{
\S4method{cbind}{lazyMatrix}(..., deparse.level = 1)

\S4method{rbind}{lazyMatrix}(..., deparse.level = 1)
}
\arguments{
\item{...}{some \code{lazyMatrix} objects}

\item{deparse.level}{ignored}
}
\description{
Concatenate two or more \code{lazyMatrix} objects.
}
