% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concat.R
\name{bind2-lazyMatrices}
\alias{bind2-lazyMatrices}
\alias{cbind2,lazyMatrix,missing-method}
\alias{cbind2,lazyMatrix,lazyMatrix-method}
\alias{cbind2,lazyVector,missing-method}
\alias{cbind2,lazyVector,lazyMatrix-method}
\alias{cbind2,lazyMatrix,lazyVector-method}
\alias{rbind2,lazyMatrix,missing-method}
\alias{rbind2,lazyMatrix,lazyMatrix-method}
\title{Concatenation of lazy matrices}
\usage{
\S4method{cbind2}{lazyMatrix,missing}(x, y)

\S4method{cbind2}{lazyVector,missing}(x, y)

\S4method{cbind2}{lazyMatrix,lazyMatrix}(x, y)

\S4method{cbind2}{lazyVector,lazyMatrix}(x, y)

\S4method{cbind2}{lazyMatrix,lazyVector}(x, y)

\S4method{rbind2}{lazyMatrix,missing}(x, y)

\S4method{rbind2}{lazyMatrix,lazyMatrix}(x, y)
}
\arguments{
\item{x, y}{\code{lazyMatrix} objects}
}
\description{
Concatenate two \code{lazyMatrix} objects.
}
