% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concat.R
\name{bind2-lazyMatrices}
\alias{bind2-lazyMatrices}
\alias{cbind2,lazyMatrix,missing-method}
\alias{cbind2,lazyMatrix,lazyMatrix-method}
\alias{cbind2,lazyVector,missing-method}
\alias{cbind2,lazyVector,lazyMatrix-method}
\alias{cbind2,lazyMatrix,lazyVector-method}
\alias{cbind2,lazyVector,lazyVector-method}
\alias{cbind2,lazyMatrix,numeric-method}
\alias{cbind2,numeric,lazyMatrix-method}
\alias{rbind2,lazyMatrix,missing-method}
\alias{rbind2,lazyMatrix,lazyMatrix-method}
\alias{rbind2,lazyVector,missing-method}
\alias{rbind2,lazyVector,lazyMatrix-method}
\alias{rbind2,lazyMatrix,lazyVector-method}
\alias{rbind2,lazyVector,lazyVector-method}
\alias{cbind2,lazyVector,numeric-method}
\alias{cbind2,numeric,lazyVector-method}
\alias{cbind2,lazyVector,matrix-method}
\alias{cbind2,matrix,lazyVector-method}
\alias{cbind2,lazyMatrix,matrix-method}
\alias{cbind2,matrix,lazyMatrix-method}
\alias{rbind2,lazyVector,numeric-method}
\alias{rbind2,numeric,lazyVector-method}
\alias{rbind2,lazyVector,matrix-method}
\alias{rbind2,matrix,lazyVector-method}
\alias{rbind2,lazyMatrix,matrix-method}
\alias{rbind2,matrix,lazyMatrix-method}
\alias{rbind2,lazyMatrix,numeric-method}
\alias{rbind2,numeric,lazyMatrix-method}
\title{Concatenation of lazy matrices}
\usage{
\S4method{cbind2}{lazyMatrix,missing}(x, y)

\S4method{cbind2}{lazyVector,missing}(x, y)

\S4method{cbind2}{lazyMatrix,lazyMatrix}(x, y)

\S4method{cbind2}{lazyVector,lazyMatrix}(x, y)

\S4method{cbind2}{lazyMatrix,lazyVector}(x, y)

\S4method{cbind2}{lazyVector,lazyVector}(x, y)

\S4method{cbind2}{lazyVector,numeric}(x, y)

\S4method{cbind2}{numeric,lazyVector}(x, y)

\S4method{cbind2}{lazyVector,matrix}(x, y)

\S4method{cbind2}{matrix,lazyVector}(x, y)

\S4method{cbind2}{matrix,lazyMatrix}(x, y)

\S4method{cbind2}{lazyMatrix,numeric}(x, y)

\S4method{cbind2}{numeric,lazyMatrix}(x, y)

\S4method{rbind2}{lazyMatrix,missing}(x, y)

\S4method{rbind2}{lazyVector,missing}(x, y)

\S4method{rbind2}{lazyMatrix,lazyMatrix}(x, y)

\S4method{rbind2}{lazyVector,lazyMatrix}(x, y)

\S4method{rbind2}{lazyMatrix,lazyVector}(x, y)

\S4method{rbind2}{lazyVector,lazyVector}(x, y)

\S4method{rbind2}{lazyVector,numeric}(x, y)

\S4method{rbind2}{numeric,lazyVector}(x, y)

\S4method{rbind2}{lazyVector,matrix}(x, y)

\S4method{rbind2}{matrix,lazyVector}(x, y)

\S4method{rbind2}{lazyMatrix,matrix}(x, y)

\S4method{rbind2}{matrix,lazyMatrix}(x, y)

\S4method{rbind2}{lazyMatrix,numeric}(x, y)

\S4method{rbind2}{numeric,lazyMatrix}(x, y)
}
\arguments{
\item{x, y}{\code{lazyMatrix} or \code{lazyVector} objects}
}
\value{
A \code{lazyMatrix} object.
}
\description{
Concatenate two \code{lazyMatrix} objects.
}
