% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyMatrices.R
\docType{methods}
\name{diag}
\alias{diag}
\alias{diag,lazyMatrix-method}
\alias{diag<-,lazyMatrix,lazyVector-method}
\title{Extract/replace diagonal of a lazy matrix}
\usage{
\S4method{diag}{lazyMatrix}(x)

\S4method{diag}{lazyMatrix,lazyVector}(x) <- value
}
\arguments{
\item{x}{a square lazy matrix}

\item{value}{a lazy vector}
}
\value{
The diagonal of \code{x} as a lazy vector, or the modified matrix.
}
\description{
Extract or replace the diagonal of a square lazy matrix.
}
