% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract.R
\name{ExtractOrReplace}
\alias{ExtractOrReplace}
\alias{[,lazyVector,numeric,missing-method}
\title{Extract/replace in a lazy vector}
\usage{
\S4method{[}{lazyVector,numeric,missing}(x, i, j, drop)
}
\arguments{
\item{x}{a \code{lazyVector} object}

\item{i}{indices}

\item{j}{nothing}

\item{drop}{ignored}
}
\description{
Extract or replace elements in a lazy vector.
}
