% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyNumbers.R
\name{intervals}
\alias{intervals}
\title{Intervals for lazy numbers}
\usage{
intervals(x)
}
\arguments{
\item{x}{a \code{lazyNumber} object}
}
\value{
A numeric matrix with two columns and \code{x@length} rows.
}
\description{
For each element in a \code{lazyNumber} object, this function 
  returns an interval containing the lazy number.
}
\examples{
library(lazyNumbers)
x <- lazynb(22) / lazynb(7)
print(intervals(x), digits = 17L)
}
