% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyNumbers.R
\docType{methods}
\name{is.na}
\alias{is.na}
\alias{is.na,lazyVector-method}
\alias{is.na,lazyMatrix-method}
\title{Missing lazy values}
\usage{
\S4method{is.na}{lazyVector}(x)

\S4method{is.na}{lazyMatrix}(x)
}
\arguments{
\item{x}{a lazy vector or a lazy matrix}
}
\value{
A logical vector or a logical matrix.
}
\description{
Check whether values are missing in lazy vectors and lazy 
  matrices.
}
\examples{
is.na(NA_lazy_)
is.na(lazyvec(c(1, 2, NA)))
}
