% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyNumbers.R
\name{lazyNumber}
\alias{lazyNumber}
\alias{as.lazyNumber}
\alias{lazynb}
\title{Lazy numbers}
\usage{
as.lazyNumber(x)

lazynb(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
An object of class \code{lazyNumber}.
}
\description{
Create a lazy number.
}
