% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyNumbers.R
\name{lazyResolve}
\alias{lazyResolve}
\title{Resolve lazy numbers}
\usage{
lazyResolve(x)
}
\arguments{
\item{x}{a \code{lazyVector} object}
}
\value{
A \code{lazyVector} object.
}
\description{
Resolve the lazy numbers in a lazy vector; see details.
}
